services/reward/requirements.txt
flask
python-dotenv
PyJWT
pymongo
requests